fix: resolve backend-size merge blockers for models/docs PR
Move models/docs wiring out of root_work.go and extract oversized test fixtures/helpers so make backend-size passes without changing CLI behavior.
